pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "axiom-cli"
version = "0.1.0"
description = "CLI orchestrator for the axiom ecosystem (continuum + axiom-team + axiom-skills)"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "pyyaml>=6",
]

[project.scripts]
axiom = "axiom_cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/axiom_cli"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8",
]
